Rename access key E2E spec file
